+++ bin/benchmark/index.ts
@@ -1,6 +1,6 @@
+import Benchmark from 'benchmark';
 import { readdir } from 'fs';
 import { resolve } from 'path';
-import Benchmark from 'benchmark';
 
 const suite = new Benchmark.Suite();
 
@@ -47,9 +47,9 @@ readdir(benchmarkPath, (err, files) => {
     process.exit(1);
   }
 
-  const benchmarks: Array<[string, Function]> = Object.entries(
+  const benchmarks: Array<[string, () => void]> = Object.entries(
     files
-      .map((name): { [key: string]: Function } => require(resolve(...paths, name)))
+      .map((name): { [key: string]: () => void } => require(resolve(...paths, name)))
       .reduce((allBenchmarks, fileBenchmarks) => Object.assign(allBenchmarks, fileBenchmarks), {}),
   );