+++ packages/rich-text-links/jest.config.js
@@ -1,8 +1,7 @@
+const getBaseConfig = require('../../baseJestConfig');
+const package = require('./package.json');
+const packageName = package.name.split('@contentful/')[1];
+
 module.exports = {
-  collectCoverage: true,
-  testPathIgnorePatterns: ['/dist/'],
-  transform: {
-    '^.+\\.tsx?$': 'ts-jest',
-  },
-  testMatch: ['**/*.test.[jt]s?(x)'],
+  ...getBaseConfig(packageName),
 };